pyproject.toml
[tool.poetry]
name = "tldr-bot"
version = "0.1.0"
description = ""
authors = ["Ramshreyas Rao <ramshreyas.rao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.1.1"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
sqlmodel = "^0.0.16"
psycopg2-binary = "^2.9" 
openai = "^1.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
